refactor: cards layout
